chore(Table): rename files to avoid conflicts with other files with the same name
