+++ src/components/Table/base/index.ts
@@ -1,7 +1,7 @@
 export { default as Header } from './Header'
 export { default as HeaderRow } from './HeaderRow'
 export { default as HeaderCell } from './HeaderCell'
-export { default as Body } from './Body'
-export { default as Row } from './Row'
-export { default as Cell } from './Cell'
+export { default as Body } from './TableBody'
+export { default as Row } from './TableRow'
+export { default as Cell } from './TableCell'
 export { default as Accordion } from './Accordion'
